% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_compensation_from_fcs}
\alias{apply_compensation_from_fcs}
\title{Apply compensation from spill matrices in provided in .fcs files.
Compensation should be applied before transforming data and before renaming or filtering channels.}
\usage{
apply_compensation_from_fcs(fs)
}
\arguments{
\item{fs}{A flowSet, flowFrame, cytoset, or cytoframe object}
}
\value{
The flowSet with compensation applied.
}
\description{
Apply compensation from spill matrices in provided in .fcs files.
Compensation should be applied before transforming data and before renaming or filtering channels.
}
