% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_channel_limits}
\alias{calculate_channel_limits}
\title{Estimate reasonable channel limits for a gate in a Gatingset.}
\usage{
calculate_channel_limits(gs, gateId, limits = NULL)
}
\arguments{
\item{gs}{A GatingSet object}

\item{gateId}{Character name of a gate}

\item{limits}{A numeric vector of the quantiles (min, max) to use to estimate the limits. If not provided, the data range is used instead (default: data range)}
}
\value{
A named list where each name is a gate and each element is the corresponding limits for the channel(s) of that gate
}
\description{
The values are calculated from all samples in the GatingSet, filtered by the gate's
parent gate. Intended use if for setting consistent limits for all samples when plotting.
}
