% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downsampleFrames}
\alias{downsampleFrames}
\title{Randomly sample from frames in a set, returning n events per frame.
If the number of events in the frame < n, all events are retained.}
\usage{
downsampleFrames(fs, n = 25000)
}
\arguments{
\item{fs}{Flowset or cytoset object}

\item{n}{Number of events to sample (default: 25000)}
}
\value{
The downsampled set
}
\description{
Randomly sample from frames in a set, returning n events per frame.
If the number of events in the frame < n, all events are retained.
}
