% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_channel_limits}
\alias{estimate_channel_limits}
\title{Estimate reasonable limits for each channel of a gate in a Gatingset.}
\usage{
estimate_channel_limits(gs, limits = NULL)
}
\arguments{
\item{gs}{A GatingSet object}

\item{limits}{A numeric vector of the quantiles (min, max) to use to estimate the limits. If not provided, the data range is used instead (default: data range)}
}
\value{
A named list where each name is a gate and each element is the corresponding limits for the channel(s) of that gate
}
\description{
The values are calculated for the entire dataset, allowing for consistent limits for plotting.
}
