% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{estimateLogicleGS}
\alias{estimateLogicleGS}
\title{Estimate the logicle transform parameters for a GatingSet. Uses all the samples
within the GatingSet and returns a \link[flowWorkspace:transformerList]{flowWorkspace::transformerList} object that
can be applied to the GatingSet using \link[flowWorkspace:transform]{flowWorkspace::transform}.}
\usage{
estimateLogicleGS(gs, channels, m = 4.5)
}
\arguments{
\item{gs}{A GatingSet object}

\item{channels}{Character vector of channels for which the transformation is to be estimated}

\item{m}{The full width of the transformed display in asymptotic decades. Should be greater than zero}
}
\value{
A \link[flowWorkspace:transformerList]{flowWorkspace::transformerList} object
}
\description{
Estimate the logicle transform parameters for a GatingSet. Uses all the samples
within the GatingSet and returns a \link[flowWorkspace:transformerList]{flowWorkspace::transformerList} object that
can be applied to the GatingSet using \link[flowWorkspace:transform]{flowWorkspace::transform}.
}
