% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_channels}
\alias{filter_channels}
\title{Filter channels (column names) in flowSet using regular expressions.}
\usage{
filter_channels(fs, keep = NULL, exclude = NULL, verbose = TRUE)
}
\arguments{
\item{fs}{The flowSet, flowFrame, cytoset, or cytoframe object}

\item{keep}{A regular expression used to select channels to keep in the flowSet.}

\item{exclude}{A regular expression used to select channels to remove from the flowSet.}

\item{verbose}{Whether to print the remaining channels (default: TRUE)}
}
\value{
The filtered flowSet.
}
\description{
Filter channels (column names) in flowSet using regular expressions.
}
\examples{
\dontrun{
# Keep only scatter and time channels
fs <- filter_channels(fs, keep="FSC|SSC|Time")
}
}
