% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_by_group}
\alias{flatten_by_group}
\title{Flatten a set by group}
\usage{
flatten_by_group(fs, groupby, rm_na = TRUE)
}
\arguments{
\item{fs}{A flowSet or cytoset object}

\item{groupby}{A column in phenoData to use to group frames}

\item{rm_na}{Whether the exclude frames with NA values in groupby column (default: TRUE)}
}
\value{
A flowSet containing flattened frame(s)
}
\description{
Groups frames by a column in phenoData then concatenates each group into a single
frame. Intended use is when the set includes replicates that you would like to combine.
}
\details{
The resultant frames will have their sampleNames attribute as the value in groupby.
All phenoData that is common to the group will be retained, and unique values
(like filenames) are discarded.

In future updates this function will return the same object class (flowSet or
cytoset) as provided. Currently, a flowSet is returned.
}
