% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_to_cytoframe}
\alias{flatten_to_cytoframe}
\title{Helper function to flatten cytoset into cytoframe}
\usage{
flatten_to_cytoframe(cs)
}
\arguments{
\item{cs}{A cytoset}
}
\value{
A cytoframe
}
\description{
Helper function to flatten cytoset into cytoframe
}
