% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_to_flowframe}
\alias{flatten_to_flowframe}
\title{Helper function to flatten flowSet into flowFrame}
\usage{
flatten_to_flowframe(fs)
}
\arguments{
\item{fs}{A flowSet}
}
\value{
A flowFrame
}
\description{
Helper function to flatten flowSet into flowFrame
}
