% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_samples_by_keyword}
\alias{get_samples_by_keyword}
\title{Searches the phenoData for provided keywords and returns the matching samples.}
\usage{
get_samples_by_keyword(
  fs,
  ...,
  use_regex = TRUE,
  ignore_case = TRUE,
  match_all = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fs}{A flowSet or related object with a phenoData slot accessible using \code{pData(fs)}}

\item{...}{Keywords to search for. Should be provided as named arguments where the name is the column to search in the phenoData and the value is the search keywords}

\item{use_regex}{Whether to search using regular expressions (default: TRUE)}

\item{ignore_case}{Whether the search is case sensitive. Only applies when use_regex = TRUE (default: TRUE)}

\item{match_all}{Whether to return samples that match all or any keywords (default: TRUE, must match all)}

\item{verbose}{Whether to print the samples that match}
}
\value{
If multiple matches are found, the returned object is the same class as
provided. If a single match, the object contained in the provided object. For example,
if a flowSet is provided, a flowFrame is returned.
}
\description{
Provide keywords as col="keyword", where col is the name of a column in phenoData.
Multiple keywords can be provided. When match_all = TRUE, a sample must match all the keywords to be returned.
When FALSE, samples that match any keyword are returned.
}
\examples{
\dontrun{
# Get all samples in group "FMO"
fs_subset <- get_samples_by_keyword(fs, group = "FMO")
}
}
