% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{lymphocyteGates}
\alias{lymphocyteGates}
\title{Automated gating of elliptical cell populations in 2D.}
\usage{
lymphocyteGates(
  gs,
  x = "FSC-A",
  y = "SSC-A",
  preselection = NULL,
  scale = 2.5,
  bwFac = 1.3,
  parentId = "root",
  gateId = "lymphocytes",
  ...
)
}
\arguments{
\item{gs}{A GatingSet, flowSet, or cytoset object}

\item{x}{Character providing a channel name}

\item{y}{Character providing a channel name}

\item{preselection}{A character giving a channel name in gs, or a named list of numerics specifying the initial rough preselection. The latter gets passed on to \link[flowCore:rectangleGate-class]{flowCore::rectangleGate}. If not provided, fits a regular \link[flowStats:norm2Filter-class]{flowStats::norm2Filter} (the default)}

\item{scale}{The scaleFactor parameter that gets passed on to \link[flowStats:norm2Filter-class]{flowStats::norm2Filter}}

\item{bwFac}{The bandwidth factor that gets passed on to \link[flowStats:curv1Filter-class]{flowStats::curv1Filter}}

\item{parentId}{Character specify the parent node name used to filter the data. Only used when a GatingSet is provided (by default, the 'root' node, no filtering is performed)}

\item{gateId}{Character specifying the name for the gate that is returned}

\item{...}{Additional arguments passed to \link[flowStats:lymphGate]{flowStats::lymphGate}}
}
\value{
Returns a list of \link[flowCore:ellipsoidGate-class]{flowCore::ellipsoidGate} objects to use for filtering
}
\description{
A convenience function for \link[flowStats:lymphGate]{flowStats::lymphGate} that constructs a gate for
each sample within a GatingSet, flowSet, or cytoset and returns a list of gates.
Additionally, when a GatingSet is provided, the gate can be constructed after
filtering by an existing gate.
}
\examples{
\dontrun{
# Construct gates after filtering by existing "Singlets" gate
gates <- lymphocyteGates(gs, x = "FSC-A", y = "SSC-A", parentId = "Singlets", gateId = "Lymphocytes")

# Add gate to GatingSet
gs_pop_add(gs, gates, parent = "Singlets")
}
}
