% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_backgating}
\alias{plot_backgating}
\title{Backgate a specified gate on all ancestor populations}
\usage{
plot_backgating(
  gs,
  gate,
  output_dir = NULL,
  output_type = "pdf",
  ncol = NULL,
  nrow = NULL,
  limits = "auto",
  labels = "channel",
  width = 27.94,
  height = 21.59,
  pos_colour = "red",
  neg_colour = "lightgray",
  alpha = 0.5,
  size = 0.3,
  remove_axis_ticks = FALSE,
  max_return = 3,
  ...
)
}
\arguments{
\item{gs}{A GatingSet object}

\item{gate}{The name of the gate used for backgating}

\item{output_dir}{Character string specifying the directory where the files will be saved}

\item{output_type}{Character string specifying the file type. Any of 'pdf', 'svg', 'png', 'tiff', or 'bmp' are supported (default: 'pdf')}

\item{ncol}{Integer specifying the number of columns in the plot grid (default: NULL)}

\item{nrow}{Integer specifying the number of rows in the plot grid (default: NULL)}

\item{limits}{Plot limits "auto", "data", "instrument" (default: "auto")}

\item{labels}{Axis labels "channel", "marker", or "both" (default: "channel")}

\item{width}{Numeric value for plot width in cm (default: 27.94, letter size)}

\item{height}{Numeric value for plot height in cm (default: 21.59, letter size)}

\item{pos_colour}{Colour of events within 'gate' (default: red)}

\item{neg_colour}{Colour of events outside 'gate' (default: lightgray)}

\item{alpha}{Alpha transparency (default: 0.5)}

\item{size}{Size of dots in scatter plots (default: 0.3)}

\item{remove_axis_ticks}{TRUE or FALSE to remove axis ticks and tick labels (numbers) for cleaner plots (default: FALSE)}

\item{max_return}{When output_dir is not provided, the generated plots will be returned and printed to the console. If the number of samples > max_return, an error will be raised. This is to prevent an excessively large number of plots being returned which may use a lot of memory (default: 3)}

\item{...}{Additional arguments passed to \code{\link[=plot_theme]{plot_theme()}}}
}
\value{
Nothing is returned when output_dir is provided. Files are saved in specified directory
}
\description{
This function generates plots for each sample in a GatingSet and saves them
as individual files.
}
\examples{
\dontrun{
library(flowWorkspace)
library(flowUtilities)

# Load a GatingSet
gs <- load_gs("path/to/gatingset")

# Plot with 3 rows
plot_backgating(gs, output_dir = "./plots", nrow = 3)
}
}
