% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gating_hierarchy}
\alias{plot_gating_hierarchy}
\title{Plot the gating hierarchy for each sample in a GatingSet}
\usage{
plot_gating_hierarchy(
  gs,
  output_dir,
  output_type = "pdf",
  ncol = NULL,
  nrow = NULL,
  limits = "data",
  width = 27.94,
  height = 21.59,
  font_size = 8,
  n_bins = 128,
  density_fill = "gray",
  density_alpha = 0.5,
  stat_loc = "gate",
  stat_digits = 1,
  stat_pos = 0.5,
  remove_axis_ticks = FALSE
)
}
\arguments{
\item{gs}{A GatingSet object containing flow cytometry data with gates}

\item{output_dir}{Character string specifying the directory where the files will be saved}

\item{output_type}{Character string specifying the file type. Any of 'pdf', 'svg', 'png', 'tiff', or 'bmp' are supported (default: 'pdf')}

\item{ncol}{Integer specifying the number of columns in the plot grid (default: NULL)}

\item{nrow}{Integer specifying the number of rows in the plot grid (default: NULL)}

\item{width}{Numeric value for plot width in cm (default: 27.94, letter size)}

\item{height}{Numeric value for plot height in cm (default: 21.59, letter size)}

\item{font_size}{Font size in points (default: 8)}

\item{n_bins}{Number of hex bins for 2D scatter plots (default: 128)}

\item{density_fill}{Color of density plots (default: gray)}

\item{density_alpha}{Alpha transparency of density plots (default: 0.5)}

\item{stat_loc}{Position of gate statistics on plot (default: "gate")}

\item{stat_digits}{Number of decimal points to plot for gate statistics (default: 1)}

\item{stat_pos}{Fine-tunes gate statistics position on plot (default: 0.5)}

\item{remove_axis_ticks}{TRUE or FALSE to remove axis ticks and tick labels (numbering) for cleaner plots (default: FALSE)}
}
\value{
Nothing is returned. Files are saved in specified directory
}
\description{
This function generates plots of the entire gating hierarchy for each sample
in a GatingSet and saves them as individual PDF files.
}
\examples{
\dontrun{
library(flowWorkspace)
library(ggcyto)

# Load a GatingSet
gs <- load_gs("path/to/gatingset")

# Plot with 3 rows
plot_gating_hierarchy(gs, output_dir = "./plots", nrow = 1)
}

}
