% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_grid}
\alias{plot_grid}
\title{Plot multiple samples in a GatingSet in a grid.}
\usage{
plot_grid(
  gs,
  x,
  y = NULL,
  gate = NULL,
  limits = "auto",
  subset = "root",
  facet_by = NULL,
  labels = "channel",
  n_bins = 128,
  ncol = 3,
  fill_density = "gray",
  fill_alpha = 0.5,
  stat_loc = "gate",
  stat_digits = 1,
  stat_pos = 0.5,
  ...
)
}
\arguments{
\item{gs}{A GatingSet object}

\item{x}{The channel to plot along the x-axis}

\item{y}{The channel to plot along the y-axis. If not specified, a 1D density histogram is plotted}

\item{gate}{A gate or gate name to plot (default: no gates plotted)}

\item{limits}{Plot limits: "auto", "data", "instrument" (default: "auto")}

\item{subset}{A gate to filter data before plotting (default: "root)}

\item{facet_by}{A column in phenoData to use to facet plots}

\item{labels}{Axis labels "channel", "marker", or "both" (default: "channel")}

\item{n_bins}{Number of hexagon bins for 2D plots (default: 128)}

\item{ncol}{Number of columns for plot grid (default: 3)}

\item{fill_density}{Fill color for 1D density histograms (default: "gray"}

\item{fill_alpha}{Transparency for 1D density histograms (default: 0.5)}

\item{stat_loc}{Position of gate statistics on plot (default: "gate")}

\item{stat_digits}{Number of decimal points to plot for gate statistics (default: 1)}

\item{stat_pos}{Fine-tunes gate statistics position on plot (default: 0.5)}

\item{...}{Additional arguments passed to \code{\link[=plot_theme]{plot_theme()}}}
}
\value{
A ggcyto or ggplot2 plot object
}
\description{
Plot multiple samples in a GatingSet in a grid.
}
