% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{quadGateFromCtrls}
\alias{quadGateFromCtrls}
\title{Constructs a quadrant gate after estimating the quantiles from controls.
Requires the phenoData to contain a column 'ctrlId', containing names of control
samples that match the channel names. This allows the function to select the correct
control for calculating the quantile. The controls are usually fluorescent minus one (FMO).}
\usage{
quadGateFromCtrls(
  gs,
  x,
  y,
  probs = 0.99,
  ctrlId = "fmo",
  parentId = "root",
  gateId = "quadGate"
)
}
\arguments{
\item{gs}{A GatingSet object}

\item{x}{Character providing a channel name}

\item{y}{Character providing a channel name}

\item{probs}{Numeric probability used to estimate the quantile (default: 0.99)}

\item{ctrlId}{Character name of column in phenoData matching control samples to channel names (default: "fmo")}

\item{parentId}{Character specify the parent node name used to filter the data (by default, the 'root' node, no filtering is performed)}

\item{gateId}{Character specifying the name for the gate that is returned}
}
\value{
A list containing a \link[flowCore:quadGate-class]{flowCore::quadGate} object accessible via $gate and a character vector of quadrant names accessible via $names.
}
\description{
Constructs a quadrant gate after estimating the quantiles from controls.
Requires the phenoData to contain a column 'ctrlId', containing names of control
samples that match the channel names. This allows the function to select the correct
control for calculating the quantile. The controls are usually fluorescent minus one (FMO).
}
\examples{
\dontrun{
# Construct gates after filtering by existing "Live" gate
gate <- quadGateFromCtrls(gs, x = "CD4", y = "CD8", parentId = "Live", gateId = "T cells")

# Add gate to GatingSet
gs_pop_add(gs, gate$gate, parent = "Singlets", names = gate$names)
}
}
