% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_filenames_to_pdata}
\alias{read_filenames_to_pdata}
\title{Extract experiment metadata from .fcs filenames and add it to phenoData.}
\usage{
read_filenames_to_pdata(fs, ...)
}
\arguments{
\item{fs}{A flowSet or GatingSet object}

\item{...}{Regular expression patterns to extract. Provide as id="regex pattern" where "id" will be added as a column in phenoData with the returned metadata as values in that column}
}
\value{
The flowSet or GatingSet with the metadata added. Can view added metadata by calling pData(fs).
}
\description{
This function uses regular expressions to find patterns in the filenames.
}
\examples{
\dontrun{
# Add a group column identifying samples and controls 
fs <- read_filenames_to_pdata(fs, group = "(?i)scc|fmo|unstained|sample")                              
}
}
