% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_names_to_pdata}
\alias{read_names_to_pdata}
\title{Extract metadata from the "names" column in phenoData and add it as a new
phenoData column.}
\usage{
read_names_to_pdata(fs, ...)
}
\arguments{
\item{fs}{A flowSet or related object with a phenoData slot accessible using \code{pData(fs)}}

\item{...}{Regular expression patterns to extract. Provide as \code{id="regex pattern"} where "id" will be added as a column in phenoData with the returned metadata as values in that column}
}
\value{
The object with the metadata added. Can view added metadata using \code{pData(fs)}.
}
\description{
This function uses regular expressions to match patterns in the filenames,
extracts those matches and adds them to a new column in phenoData.
}
\examples{
\dontrun{
# Add a group column identifying samples and controls
fs <- read_names_to_pdata(fs, group = "(?i)scc|fmo|unstained|sample")
}
}
