% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{singletsGates}
\alias{singletsGates}
\title{Constructs a singlet gate by applying a robust linear model with \link[MASS:rlm]{MASS::rlm}.}
\usage{
singletsGates(
  gs,
  area = "FSC-A",
  height = "FSC-H",
  sidescatter = NULL,
  prediction_level = 0.99,
  subsample_pct = NULL,
  wider_gate = FALSE,
  parentId = "root",
  gateId = "singlets",
  maxit = 10,
  ...
)
}
\arguments{
\item{gs}{A GatingSet, flowSet, or cytoset object}

\item{area}{Character giving the channel name that records the signal intensity as peak area}

\item{height}{Character giving the channel name that records the signal intensity as peak height}

\item{sidescatter}{UNUSED Character giving an optional channel name for the sidescatter signal (by default, ignored)}

\item{prediction_level}{A numeric value between 0 and 1 specifying the level to use for the prediction bands}

\item{subsample_pct}{A numeric value between 0 and 1 indicating the percentage of observations that should be randomly selected from x to construct the gate (by default, no subsampling is performed)}

\item{wider_gate}{Logical value. If TRUE, the prediction bands used to construct the singlet gate use the robust fitted weights, which increase prediction uncertainty, especially for large FSC-A. This leads to wider gates, which are sometimes desired}

\item{parentId}{Character specify the parent node name used to filter the data. Only used when a GatingSet is provided (by default, the 'root' node, no filtering is performed)}

\item{gateId}{Character specifying the name for the gate that is returned}

\item{maxit}{the limit on the number of IWLS iterations passed to \link[MASS:rlm]{MASS::rlm}}

\item{...}{Additional arguments passed to \link[MASS:rlm]{MASS::rlm}}
}
\value{
Returns a list of \link[flowCore:polygonGate-class]{flowCore::polygonGate} objects to use for filtering
}
\description{
A convenience function for \link[flowStats:gate_singlet]{flowStats::singletGate} that constructs a gate for
each sample within a GatingSet, flowSet, or cytoset and returns a list of gates.
Additionally, when a GatingSet is provided, the singlet gate can be constructed after filtering by an existing gate.
}
\examples{
\dontrun{
# Construct gates after filtering by existing "NonDebris" gate
gates <- singletsGates(gs, area = "FSC-A", height = "FSC-H", parentId = "NonDebris", gateId = "Singlets")

# Add gate to GatingSet
gs_pop_add(gs, gates, parent = "NonDebris")
}
}
