% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{autoQuadGate}
\alias{autoQuadGate}
\title{Constructs a quadrant gate after estimating the quantiles from controls.
Requires the phenoData to contain a column 'ctrlId', containing names of control
samples that match the channel names. This allows the function to select the correct
control for calculating the quantile. The controls are usually fluorescent minus one (FMO).}
\usage{
autoQuadGate(
  gs,
  x,
  y,
  probs = 0.99,
  ctrlId = "fmo",
  parentId = "root",
  gateId = "quadGate"
)
}
\arguments{
\item{gs}{A GatingSet object}

\item{x}{Character providing a channel name}

\item{y}{Character providing a channel name}

\item{probs}{Numeric probability used to estimate the quantile (default: 0.99)}

\item{ctrlId}{Character name of column in phenoData matching control samples to channel names (default: "fmo")}

\item{parentId}{Character specify the parent node name used to filter the data (by default, the 'root' node, no filtering is performed)}

\item{gateId}{Character specifying the name for the gate that is returned}
}
\value{
A \link[flowCore:quadGate-class]{flowCore::quadGate} object
}
\description{
Constructs a quadrant gate after estimating the quantiles from controls.
Requires the phenoData to contain a column 'ctrlId', containing names of control
samples that match the channel names. This allows the function to select the correct
control for calculating the quantile. The controls are usually fluorescent minus one (FMO).
}
\examples{
\dontrun{
# Construct gates after filtering by existing "Live" gate
gate <- autoQuadGate(gs, x = "CD4", y = "CD8", parentId = "Live")

# A quadGate becomes four rectangleGate when added to a GatingSet
# The names will be created automatically or, we can specify the names
# when adding to GatingSet (order is clockwise from upper left)
names <- c("CD4-CD8+","CD4+CD8+","CD4+CD8-","CD4-CD8-")

# Add gate to GatingSet
gs_pop_add(gs, gate, parent = "Live", names = names)
}
}
