% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_quantile}
\alias{calculate_quantile}
\title{Calculates the quantile for a given channel (column).}
\usage{
calculate_quantile(gs, channels = NULL, probs = 0.99, parentId = "root", ...)
}
\arguments{
\item{gs}{GatingSet, GatingHierarchy, flowSet, flowFrame, cytoset or cytoframe object}

\item{channels}{Character vector specifying the channels (column names) in gs to calculate the quantile for (default: all channels)}

\item{parentId}{Optional name of gate to use to filter data before calculating quantile. Only used when a GatingSet or GatingHierarchy is provided}

\item{...}{Additional arguments passed to \link[stats:quantile]{stats::quantile}}

\item{prob}{A numeric probability value between 0 and 1 or a vector of probabilities (by default, 0.99)}
}
\value{
A named list of quantile values
}
\description{
Compatible with GatingSet, GatingHierarchy, flowSet, flowFrame, cytoset, and cytoframe objects.
If a GatingSet, flowSet or cytoset is provided containing multiple samples, all data is
concatenated and the quantile calculated using all of the data.
}
\examples{
\dontrun{
# Calculate quantile values for the CD4 channels using a flowSet, flowFrame, cytoset, or cytoframe
q <- calculate_quantile(fs, "CD4", prob = 0.99)

# Calculate quantile values for the CD4 channels using a GatingSet or GatingHierarchy by filtering for "Live" cells first
q <- calculate_quantile(gs, "CD4", prob = 0.99, parentId = "Live")
}
}
