% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_range}
\alias{calculate_range}
\title{Calculates the range (min, max) for a given channel (column).}
\usage{
calculate_range(gs, channels = NULL, parentId = "root")
}
\arguments{
\item{gs}{GatingSet, GatingHierarchy, flowSet, flowFrame, cytoset or cytoframe object}

\item{channels}{Character vector or list specifying the channels (column names) in gs to calculate the range for (default: all channels)}

\item{parentId}{Optional name of gate to use to filter data before calculating quantile. Only used when a GatingSet or GatingHierarchy is provided}
}
\value{
A named list of range values
}
\description{
Compatible with GatingSet, GatingHierarchy, flowSet, flowFrame, cytoset, and cytoframe objects.
If a GatingSet, flowSet or cytoset is provided containing multiple samples, all data is
concatenated and the quantile calculated using all of the data.
}
\examples{
\dontrun{
# Calculate range values for the CD4 channels using a flowSet
r <- calculate_range(fs, "CD4")

# Calculate range values for the CD4 channels using a GatingSet
# by filtering for "Live" cells first
r <- calculate_range(gs, "CD4", parentId = "Live")
}
}
