% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{diagPolygonGate}
\alias{diagPolygonGate}
\title{Constructs a 6 vertex diagonal polygon gate. Useful for gating singlets using
FSC-A x FSC-H or gating CD3+TCRab+ T-cells.}
\usage{
diagPolygonGate(
  x,
  y,
  width,
  height,
  origin = c(0, 0),
  diag_width = 0.5,
  gateId = "PolygonGate"
)
}
\arguments{
\item{x}{The name of the channel along the width (x-axis)}

\item{y}{The name of the channel along the height (y-axis)}

\item{width}{The total width of the polygon}

\item{height}{The total height of the polygon}

\item{origin}{A numeric vector of length 2 defining the position of the bottom left corner of the gate (the default is c(0,0))}

\item{diag_width}{The width between the two diagonal lines (the default is 0.5)}

\item{gateId}{Character specifying the name for the gate that is returned}
}
\value{
Returns a \link[flowCore:polygonGate-class]{flowCore::polygonGate} object to use for filtering
}
\description{
The polygon has a maximum width defined by the width argument, a maximum height
defined by the height argument, and a width between diagonal lines defined by diag_width.
The bottom left corner is positioned at the origin argument.
}
\examples{
\dontrun{
# Construct a singlets gate
gates <- diagPolygonGate(gs,
                         x = "FSC-A",
                         y = "FSC-H",
                         width = 3e6,
                         height = 3.5e6,
                         diag_width = 7e5,
                         gateId = "Singlets")

# Add gate to GatingSet
gs_pop_add(gs, gates, parent = "root")
}
}
