% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_channels}
\alias{filter_channels}
\title{Filter channels in set using regular expressions.}
\usage{
filter_channels(
  fs,
  keep = NULL,
  exclude = NULL,
  filter_scatter = FALSE,
  ignore_case = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fs}{The flowSet, flowFrame, cytoset, or cytoframe object}

\item{keep}{A regular expression used to select channels to keep in the set}

\item{exclude}{A regular expression used to select channels to remove from the set}

\item{filter_scatter}{Whether to allow filtering of scatter channels. When TRUE, the FSC/SSC channels can be removed from the set (default: FALSE)}

\item{ignore_case}{Whether to ignore case in the regular expression such that "fl1-a" matches "FL1-A" etc (default: TRUE)}

\item{verbose}{Whether to print the remaining channels (default: TRUE)}
}
\value{
The filtered flowSet.
}
\description{
Filter channels in set using regular expressions.
}
\examples{
\dontrun{
# Keep only scatter and time channels
fs <- filter_channels(fs, keep="FSC|SSC|Time")
}
}
