% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_marker_names}
\alias{get_marker_names}
\title{Returns the marker names stored in .fcs files.}
\usage{
get_marker_names(
  fs,
  ...,
  sep = "",
  default_name = NULL,
  allow_partial_match = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fs}{A flowSet containing multiple .fcs files}

\item{...}{Any number of regular expression patterns (strings) to extract from marker names.}

\item{sep}{The separator to use when concatenating strings (default: "")}

\item{default_name}{Optional name for marker names where no patterns match (default: none)}

\item{allow_partial_match}{Whether to use marker names when not all regular expression patterns match (default: TRUE)}

\item{verbose}{Whether to print updated markers (default: TRUE)}
}
\value{
A list of marker names for their corresponding channels. This can be applied to fs
using \code{\link[=set_channels_from_markers]{set_channels_from_markers()}}.
}
\description{
The marker names are typically user-defined when the .fcs files are created
and can include descriptive names like "CD45 Y585-PE-A". You can use markernames(fs)
to view the current marker names. Does not alter scatter (FCS-A, SSC-A, etc.)
or time channels.
}
\details{
Optionally, regular expression(s) may be used to extract substrings from marker names.
When multiple patterns are provided, the substrings will be concatenated together in the order provided.
This can be used to avoid duplicate channel names. For example, when both area and height are recorded for a given marker, it may be useful to distinguish them using the pattern "-A|-H".
If the provided pattern does not match, the full marker name is returned unless partial matches are permitted.

If desired, a default name can be provided for marker names without a match. This can be used to label unused channels
}
\examples{
\dontrun{
# Get marker names
markers <- get_marker_names(fs)
}
}
