% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_samples_by_keyword}
\alias{get_samples_by_keyword}
\title{Searches the phenoData for provided keywords and returns the matching samples.
If a match is found, returns the associated flowFrame when a flowSet is provided, or a GatingHierarchy when a GatingSet is provided.
If multiple matches are found, a flowSet/GatingSet is returned containing those matches. If no matches, the original data is returned.}
\usage{
get_samples_by_keyword(
  fs,
  ...,
  use_regex = TRUE,
  ignore_case = TRUE,
  match_all = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fs}{A flowSet or GatingSet}

\item{...}{Keywords to search for. Should be provided as named arguments where the name is the column to search in the phenoData and the value is the search keywords}

\item{use_regex}{Whether to search using regular expressions (default: TRUE)}

\item{ignore_case}{Whether the search is case sensitive. Only applies when use_regex = TRUE (default: TRUE)}

\item{match_all}{Whether to return samples that match all or any keywords (default: TRUE, must match all)}

\item{verbose}{Whether to print the samples that match}
}
\value{
A flowFrame, flowSet, GatingSet, or GatingHierarchy depending on what was provided
}
\description{
Provide keywords as col="keyword", where col is the name of a column in phenoData.
Multiple keywords can be provided. When match_all = TRUE, a sample must match all the keywords to be returned.
When FALSE, samples that match any keyword are returned.
}
\examples{
\dontrun{
# Get all samples in group "FMO"
fs_subset <- get_samples_by_keyword(fs, group = "FMO")
}

}
