% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_spill_matrix}
\alias{get_spill_matrix}
\title{Returns the spill matrix contained in a flowFrame.}
\usage{
get_spill_matrix(ff)
}
\arguments{
\item{ff}{A flowFrame or cytoframe object}
}
\value{
The spill matrix for the flowFrame, if found, or else NULL
}
\description{
If none is present, returns NULL.
}
