% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_grid}
\alias{plot_grid}
\title{Plot a grid of samples in a GatingSet.}
\usage{
plot_grid(
  gs,
  x,
  y = NULL,
  gate = NULL,
  limits = "data",
  subset = "root",
  facet_by = NULL,
  n_bins = 128,
  n_cols = 3,
  fill_density = "gray",
  fill_alpha = 0.5,
  stat_loc = "gate",
  stat_digits = 1,
  stat_pos = 0.5
)
}
\arguments{
\item{gs}{A GatingSet object}

\item{x}{The channel to plot along the x-axis}

\item{y}{The channel to plot along the y-axis. If not specified, a 1D density histogram is plotted}

\item{gate}{A gate or gate name to plot (default: no gates plotted)}

\item{limits}{Plot limits: "auto", "data", "instrument", a list("x"=c(min, max), "y"=c(min, max)), a numeric vector of quantile values}

\item{subset}{A gate to filter data before plotting (default: "root)}

\item{facet_by}{A column in phenoData to use to facet plots}

\item{n_bins}{Number of hexagon bins for 2D plots (default: 128)}

\item{n_cols}{Number of columns for plot grid (default: 3)}

\item{fill_density}{Fill color for 1D density histograms (default: "gray"}

\item{fill_alpha}{Transparency for 1D density histograms (default: 0.5)}

\item{stat_loc}{Position of gate statistics on plot (default: "gate")}

\item{stat_digits}{Number of decimal points to plot for gate statistics (default: 1)}

\item{stat_pos}{Fine-tunes gate statistics position on plot (default: 0.5)}
}
\value{
A ggcyto or ggplot2 plot object
}
\description{
Plot a grid of samples in a GatingSet.
}
