% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_stacked_hist}
\alias{plot_stacked_hist}
\title{Creates a stacked density histogram for each channel provided. Does not plot gates -
this function is intended to plot channels when applying transformations.}
\usage{
plot_stacked_hist(
  gs,
  channels,
  stack_by = "name",
  subset = "root",
  alpha = 0.5
)
}
\arguments{
\item{gs}{A GatingSet object}

\item{channels}{Character names of channels in gs to plot}

\item{stack_by}{A column in phenoData to use to arrange histograms along the y-axis (default: "name")}

\item{subset}{A gate to filter data before plotting (default: "root)}

\item{alpha}{Density transparency}
}
\value{
A ggplot2 gtable object
}
\description{
Creates a stacked density histogram for each channel provided. Does not plot gates -
this function is intended to plot channels when applying transformations.
}
