% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_channels_from_markers}
\alias{set_channels_from_markers}
\title{Set channels from the marker names in a flowSet or GatingSet.}
\usage{
set_channels_from_markers(fs, markers = NULL, set_markers_as_channels = FALSE)
}
\arguments{
\item{fs}{A flowSet or GatingSet object}

\item{markers}{A list of marker names to set as channel names. If NULL, the stored marker names are used}

\item{set_markers_as_channels}{Whether to set the old channel names as marker names (default: FALSE)}
}
\value{
The flowSet with updated channel and marker names
}
\description{
The original channel names become the marker names. The marker names can be provided as a list, such as what is
returned from \code{\link[=get_marker_names]{get_marker_names()}}. If not provided, the marker names contained in the flowSet are used.
}
